refactor(your-story): restructure StoryComponentWrapper for better modularity, maintainability and readability
