+++ src/app/(pages)/fathers-and-figures/StoryCards/StoryCardSetWrapper.tsx
@@ -2,6 +2,8 @@
 import React, { useContext } from 'react'
 import { StoryCardSet } from '../StoryCards'
 import { MyStoriesContext } from '../YourStory/MyStoriesProvider'
+import ButtonClearDraft from './ButtonClearDraft'
+import ButtonQueueDraft from './ButtonQueueDraft'
 
 const StoryCardSetWrapper= () => {
 
@@ -9,11 +11,8 @@ const StoryCardSetWrapper= () => {
 
     if (!context) return null
 
-    const { myStories, setMyStories } = context
+    const { myStories } = context
 
-    const handleClearDrafts = () => {
-        setMyStories([])
-    }
 
     if (myStories.length === 0) {
         return (
@@ -29,20 +28,8 @@ const StoryCardSetWrapper= () => {
         <StoryCardSet myStories={myStories} />
         </div>
         <div className='text-container justify-between flex gap-2 mx-0 my-2'>
-            <button 
-                type='button' 
-                onClick={() => {handleClearDrafts()}} 
-                className='border-[1px] flex-1 text-xs rounded-sm border-[var(--primary-blue)] p-1 hover:scale-90 hover:shadow-[4px_4px_0px_0px_var(--primary-blue)] transition duration-300 ease-in-out'
-            >
-                Start with a clean slate? Clear all drafts!
-            </button> 
-            <button 
-                type='button' 
-                onClick={() => {alert('Hang on, Sumit is still cooking this feature!')}} 
-                className='border-[1px] flex-1 text-xs rounded-sm border-[var(--primary-blue)] p-1 hover:scale-90 hover:shadow-[4px_4px_0px_0px_var(--primary-blue)] transition duration-300 ease-in-out'
-            >
-                Queue up your story for publishing below!
-            </button> 
+            <ButtonClearDraft />
+            <ButtonQueueDraft />
         </div>
     </>
   )
